cleanup the proc_tools library
fix build warnings
